Move SearchScope
